#41 時間の登録は30分単位で行えるようにする

+++ backend/app/models/time_model.py
@@ -4,19 +4,19 @@ from typing import Union
 
 class RegisterActivities(BaseModel):
     date: str
-    target_time: Union[int, str]
-    actual_time: Union[int, str]
+    target_time: Union[float, str]
+    actual_time: Union[float, str]
     is_achieved: Union[bool, str]
     message: str
 
 
 class TargetTimeIn(BaseModel):
-    target_time: int
+    target_time: float
 
 
 class ActualTimeIn(BaseModel):
-    actual_time: int
+    actual_time: float
 
 
 class RegisterSalary(BaseModel):
-    salary: int
+    salary: float
